Make it more clean. put onApproveTimeEntries at useSuperAdminHomePageContext
