+++ src/context/SuperAdminHomePageContext/SuperAdminHomePageContext.tsx
@@ -8,16 +8,19 @@ import {
 } from "../../utilily/types";
 import { useSuperAdminHomePageContext } from "./useSuperAdminHomePageContext";
 import { getAllUnconfirmedTimeEntries } from "../../firebase-functions/getTS/get";
-import { confirmTimeEntry } from "../../firebase-functions/updateTS/update";
-import { filter, isEmpty, reject, remove, set } from "lodash";
+// import { confirmTimeEntry } from "../../firebase-functions/updateTS/update";
+// import { filter, isEmpty, reject, remove, set } from "lodash";
 
 type SuperAdminHomePageContextType = {
   allUsersWithUnconfirmedTimeEntries: Array<UserAndUnapprovedTimeEntriesType>;
   getAllUserAndUnapprovedTimeEntries: () => void;
-  onApproveTimeEntries: (
-    timeEntries: Array<TimeEntry["id"]>,
-    approvedBy: User["id"],
+  setAllUsersWithUnconfirmedTimeEntries: (
+    users: UserAndUnapprovedTimeEntriesType[],
   ) => void;
+  // onApproveTimeEntries: (
+  //   timeEntries: Array<TimeEntry["id"]>,
+  //   approvedBy: User["id"],
+  // ) => void;
 };
 
 const SuperAdminHomePageContext = React.createContext<
@@ -47,33 +50,34 @@ export const SuperAdminHomePageContextProvider: React.FC<{
       setAllUsersWithUnconfirmedTimeEntries(usersAndUnconfirmedTimeEntries);
   };
 
-  const onApproveTimeEntries = (
-    timeEntries: Array<TimeEntry["id"]>,
-    approvedBy: User["id"],
-  ) => {
-    timeEntries.map(
-      async (timeEntryID) =>
-        await confirmTimeEntry(timeEntryID, approvedBy).then(() => {
-          setAllUsersWithUnconfirmedTimeEntries((prevUsers) => {
-            return prevUsers
-              .map((user) => ({
-                ...user,
-                unapprovedTimeEntries: reject(user.unapprovedTimeEntries, {
-                  id: timeEntryID,
-                }) as TimeEntry[],
-              }))
-              .filter((user) => user.unapprovedTimeEntries.length > 0);
-          });
-        }),
-    );
-  };
+  // const onApproveTimeEntries = (
+  //   timeEntries: Array<TimeEntry["id"]>,
+  //   approvedBy: User["id"],
+  // ) => {
+  //   timeEntries.map(
+  //     async (timeEntryID) =>
+  //       await confirmTimeEntry(timeEntryID, approvedBy).then(() => {
+  //         setAllUsersWithUnconfirmedTimeEntries((prevUsers) => {
+  //           return prevUsers
+  //             .map((user) => ({
+  //               ...user,
+  //               unapprovedTimeEntries: reject(user.unapprovedTimeEntries, {
+  //                 id: timeEntryID,
+  //               }) as TimeEntry[],
+  //             }))
+  //             .filter((user) => user.unapprovedTimeEntries.length > 0);
+  //         });
+  //       }),
+  //   );
+  // };
 
   return (
     <SuperAdminHomePageContext.Provider
       value={{
         allUsersWithUnconfirmedTimeEntries,
         getAllUserAndUnapprovedTimeEntries,
-        onApproveTimeEntries,
+        setAllUsersWithUnconfirmedTimeEntries,
+        // onApproveTimeEntries,
       }}
     >
       {children}
